RF: Refactoring the bot and layout creation methods.
